Update README.md
There was a typo in the path linked in #2 of Week 03 that lead to a 404 File Not Found page when clicked. (the typo was that "master" was spelled "mastAer") Went in in fixed the typo so that it will hopefully lead students to the correct "ArcGIS_Online_FieldMaps.md" file.